backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
finnhub-python==2.4.19
yfinance==0.2.28
httpx==0.25.2
uvicorn==0.24.0
